Fix Test.get_files regex matching
Match regex on string instead of matching string on regex. So there was an
error in the code, which was harder for me to spot since I was matching with a
shell glob, hence it was hard to noticed what went wrong. This issue has been
fixed and a test has been added to resolve this error. I however also updated
the comment banner for future me.