Create test for verbosity options
This pull request fixs the failing test that was added in https://github.com/DigitPaint/roger/issues/12.
The main problem with this test is that we rely on the Project to hold the passed class options,
this however is not the case for the helpers which initiliaze their own project.
What I did now is, I create a Cli::Base manualy and pass a --verbose on init time.
After I added my tests, I invoke this initialised CLI. Which passes the --verbose
to Project in its initialized, which is skipped in the helpers.

I guess we could redesign the helpers, seperate it as it tests several
concerns or just don't test it at all. For now, I've added this one, for your consideration.
It is however safe to merge, imo.