Setup test coverage with integration to code climate
For this to work I extracted all test_unit/mocha/helpers requires
to one file, this one gets required in all the test nows, for here we
can start the test coverage.

We are using code_climate when ENV["CODECLIMATE_REPO_TOKEN"] is set,
because when running code_climate during development it will also push
coverage which is less optimal.