Resolve to file when dirname and filename conflict
By convention resolving a directory like '/elements/' will resolve
to the index.html in this directory `=> 'elements/index.html'` when
available. This is done by checking that the passed in `name` in
`find_template` is a directory on the line altered in this commit. It
does not take into account that this name can also be a file (when
passed in without extension). This is an issue because most template
render call work without extension and traverse this same code:

```
  <%= partial 'elements/footer' %>
```

Without this commit, just for the fact there is a directory called
`footer` it will not resolve to the footer.html.erb.

```
- elements
  - footer/
     - disclaimer
  - footer.html.erb
```

This commits changes the way index.xyz handling is done. It will first search for possible files and as a last resort it will check to see if there is a directory with the same name, append index to the name and finally will call the find_template method again with the appended index.